back_geba/requirements.txt
asgiref==3.7.2
build==1.0.3
CacheControl==0.13.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
crashtest==0.4.1
distlib==0.3.7
Django==5.0
django-cors-headers==4.3.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dulwich==0.21.7
fastjsonschema==2.19.0
filelock==3.13.1
idna==3.6
importlib-metadata==7.0.0
installer==0.7.0
jaraco.classes==3.3.0
keyring==24.3.0
more-itertools==10.1.0
msgpack==1.0.7
packaging==23.2
pexpect==4.9.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycparser==2.21
PyJWT==2.8.0
pyproject_hooks==1.0.0
pytz==2023.3.post1
rapidfuzz==3.5.2
requests==2.31.0
requests-toolbelt==1.0.0
shellingham==1.5.4
sqlparse==0.4.4
tomlkit==0.12.3
trove-classifiers==2023.11.29
urllib3==2.1.0
virtualenv==20.25.0
xattr==0.10.1
zipp==3.17.0
